pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "poe-crawler"
version = "1.0.0"
description = "A crawler for Poe bot information and pricing"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Poe Crawler Author"}
]

dependencies = [
    "requests==2.31.0",
    "python-dotenv==1.0.1",
    "beautifulsoup4==4.12.3",
    "jinja2==3.1.3",
    "markdown==3.5.2",
    "tqdm==4.66.2",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "src",
    "view_html.py",
    "README.md",
    "pyproject.toml",
]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[project.scripts]
poe-crawler = "src.main:main"
view-html = "view_html:main"
